separated SearchHeader component, minor refactor
